+++ src/app/code/community/Shopgate/Cloudapi/Model/Api2/Customers/Rest/Customer/V2.php
@@ -20,27 +20,34 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
 
-class Shopgate_Cloudapi_Model_Api2_Auth_Token_Rest_Guest_V2 extends Shopgate_Cloudapi_Model_Api2_Auth_Rest
+class Shopgate_Cloudapi_Model_Api2_Customers_Rest_Customer_V2 extends Shopgate_Cloudapi_Model_Api2_Customers_Rest
 {
     /** @noinspection PhpHierarchyChecksInspection */
     /**
-     * Creates access_tokens based on grant_type in body
+     * Retrieve customer data
      *
-     * @param array $data - incoming request parameters from body
-     *
-     * @return array
-     * @throws LogicException
-     * @throws InvalidArgumentException
-     * @throws Zend_Controller_Response_Exception
      * @throws Exception
      */
-    public function _create(array $data)
+    protected function _retrieve()
     {
-        $server = Mage::getModel('shopgate_cloudapi/oAuth2_server')->initialize($this->_getStore());
-        /** @var \OAuth2\Response $response */
-        $response = $server->handleTokenRequest(OAuth2\Request::createFromGlobals());
-        Mage::helper('shopgate_cloudapi/oAuth2_response')->translate($response);
+        $customerData = Mage::getModel('customer/customer_api')->info($this->getApiUser()->getUserId());
 
-        return $response->getParameters();
+        return $this->filterOutData($customerData);
+    }
+
+    /**
+     * Removes sensitive data from retrieval
+     *
+     * @todo-sg: should use the native filter in Shopgate_Cloudapi_Model_Api2_Resource::dispatch
+     *
+     * @param array $data
+     *
+     * @return array
+     */
+    private function filterOutData(array $data)
+    {
+        $excludeKeys = array('password_hash');
+
+        return array_diff_key($data, array_flip($excludeKeys));
     }
 }